requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Flask-Migrate==2.5.3
Flask-Login==0.5.0
Flask-Caching==1.9.0
gunicorn==20.0.4
